Aggregated statements

+++ backend/src/main/java/com/example/demo/domain/RankedItem.java
@@ -1,15 +1,14 @@
 package com.example.demo.domain;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
+import jakarta.validation.constraints.NotNull;
+import lombok.*;
 
 @AllArgsConstructor
 @Data
 @Builder
 public class RankedItem {
     private int itemId;
-    private int units;
+    private double units;
     private double totalIncome;
     private String name;
 }
